server/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
pydantic==2.9.2
requests==2.32.5
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous>=2.1.2,<3.0.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4